+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 5.4.0 (unreleased)
 
 * Deprecated several commands in favor of using policy.
+* Add `policy` subcommand for `ldap-sync`, get rid if `jobs` and `now` subcommands.
 
 # 5.3.0